refactor: change code in process test suite
