+++ packages/untp-test-suite/src/templates/index.ts
@@ -1,0 +1,2 @@
+export * from './getTemplateName.js';
+export * from './generateFinalMessage.js';
